requirements.txt
# Core Django
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3

# Export
openpyxl==3.1.2
